SceneGraph: A revision for ILP32 systems of the changes in r87 ("ambiguous call to overloaded function").
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@88 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/SceneGraph/BezierPatchNode.cpp
@@ -194,8 +194,8 @@ void BezierPatchNodeT::WriteTo(std::ostream& OutFile, aux::PoolT& Pool) const
     aux::Write(OutFile, aux::cnc32(SizeX));
     aux::Write(OutFile, aux::cnc32(SizeY));
 
-    aux::Write(OutFile, SubdivsHorz);
-    aux::Write(OutFile, SubdivsVert);
+    aux::Write(OutFile, (int32_t)SubdivsHorz);
+    aux::Write(OutFile, (int32_t)SubdivsVert);
  // aux::Write(OutFile, MaxError);
     assert(MaxError==400.0f);
